+++ client/src/pages/SignUp/SignUp.js
@@ -6,102 +6,53 @@ import "./SignUp.css"
 
 class SignUp extends Component {
 
-
-  constructor(props){
-    super(props)
-    this.setState = {
-      array: ['X', 'PageTwo', 'Array'],
-      string: 'XPageTwoString',
-      currentUserEmail: '',
-      currentUserName: '',
-      value: '',
-      handleSubmit: this.handleSubmit.bind(this),
-      handleChange: this.handleChange.bind(this)
-    }
-  }
-
-  componentDidMount() {
-    API.getDocuments()
-      .then((res) => { console.log(res.data) })
-      .catch((err) => console.log(err));
-
-
-    //parse it so you can grab the values using . notation
-    const idToken = JSON.parse(localStorage.getItem('okta-token-storage'));
-
-    this.setState({
-      currentUserEmail: idToken.idToken.claims.currentUserEmail,
-      currentUserName: idToken.idToken.claims.name
-    })
-  }
-
-  handleChange(event){
-    this.setState({value: event.target.value})
-  }
-
-  handleSubmit(event){
-    event.preventDefault()
-
-  }
-
   render() {
-    console.log(this.state)
-    //should get currentUserEmail & currentUserName
-
-    const { currentUserEmail, currentUserName } = this.state;
     return (
-    <div>
-      <MDBContainer className="SignForm">
-        <MDBRow>
-          <MDBCol md="6" className="ActForm">
+      <div>
+        <MDBContainer className="LogForm">
+          <MDBRow>
+            <MDBCol className="ActForm" md="6">
+              <form>
+                <p className="h5 text-center mb-4">Sign Up</p>
+                <div className="grey-text">
+                  <MDBInput
+                    label="Type your name"
+                    icon="user-alt"
+                    group
+                    type="email"
+                    validate
+                    error="wrong"
+                    success="right"
+                  />
+                  <MDBInput
+                    label="Type your email"
+                    icon="envelope"
+                    group
+                    type="email"
+                    validate
+                    error="wrong"
+                    success="right"
+                  />
+                  <MDBInput
+                    label="Type your password"
+                    icon="lock"
+                    group
+                    type="password"
+                    validate
+                  />
+                </div>
+                <div className="text-center">
+                  <Link to="/Creator">
+                    <MDBBtn >Done</MDBBtn>
+                  </Link>
+                </div>
+              </form>
+            </MDBCol>
+          </MDBRow>
+        </MDBContainer>
 
-            <form onSubmit={this.handleSubmit}>
-              <p className="h5 text-center mb-4">Sign up</p>
-              <div className="grey-text">
-                <MDBInput
-                  label="Your name"
-                  icon="user"
-                  group
-                  type="text"
-                  validate
-                  error="wrong"
-                  success="right"
-                  //maybe value ={this.state.value}
-                  value={this.state.currentUserName} onChange={this.handleChange}
-                />
-                <MDBInput
-                  label="Your email"
-                  icon="envelope"
-                  group
-                  type="email"
-                  validate
-                  error="wrong"
-                  success="right"
-                  //maybe value = {this.state.value}
-                  value={this.state.currentUserEmail}
-                  onChange={this.handleChange}
-                />
-
-                <MDBInput
-                  label="Your password"
-                  icon="lock"
-                  group
-                  type="password"
-                  validate
-                  value={this.state.value}
-                  
-                />
-        
-              </div>
-              <div className="text-center">
-                <MDBBtn type="submit" value="submit">Register</MDBBtn>
-              </div>
-            </form>
-
-          </MDBCol>
-        </MDBRow>
-      </MDBContainer>
-    </div>);
+        {/* <Footer /> */}
+      </div>);
   }
 }
